+++ baybe/searchspace.py
@@ -10,8 +10,16 @@ import numpy as np
 import pandas as pd
 import torch
 from attrs import define, field
+from botorch.utils.sampling import get_polytope_samples
 
-from baybe.constraints import _validate_constraints, Constraint, CONSTRAINTS_ORDER
+from baybe.constraints import (
+    _validate_constraints,
+    Constraint,
+    ContinuousLinearEqualityConstraint,
+    ContinuousLinearInequalityConstraint,
+    DISCRETE_CONSTRAINTS_FILTERING_ORDER,
+    DiscreteConstraint,
+)
 from baybe.parameters import (
     _validate_parameter_names,
     _validate_parameters,
@@ -26,6 +34,7 @@ from baybe.parameters import (
 from baybe.telemetry import TELEM_LABELS, telemetry_record_value
 from baybe.utils import (
     df_drop_single_value_columns,
+    DTypeFloatTorch,
     eq_dataframe,
     fuzzy_row_match,
     SerialMixin,
@@ -77,7 +86,7 @@ class SubspaceDiscrete:
     exp_rep: pd.DataFrame = field(eq=eq_dataframe)
     metadata: pd.DataFrame = field(eq=eq_dataframe)
     empty_encoding: bool = field(default=False)
-    constraints: List[Constraint] = field(factory=list)
+    constraints: List[DiscreteConstraint] = field(factory=list)
     comp_rep: pd.DataFrame = field(eq=eq_dataframe)
 
     @exp_rep.validator
@@ -163,7 +172,7 @@ class SubspaceDiscrete:
     def from_product(
         cls,
         parameters: List[DiscreteParameter],
-        constraints: Optional[List[Constraint]] = None,
+        constraints: Optional[List[DiscreteConstraint]] = None,
         empty_encoding: bool = False,
     ) -> "SubspaceDiscrete":
         """See :py:class:`baybe.searchspace.SearchSpace`."""
@@ -173,7 +182,8 @@ class SubspaceDiscrete:
         else:
             # Reorder the constraints according to their execution order
             constraints = sorted(
-                constraints, key=lambda x: CONSTRAINTS_ORDER.index(x.__class__)
+                constraints,
+                key=lambda x: DISCRETE_CONSTRAINTS_FILTERING_ORDER.index(x.__class__),
             )
 
         # Create a dataframe representing the experimental search space
@@ -267,7 +277,7 @@ class SubspaceDiscrete:
     def param_bounds_comp(self) -> torch.Tensor:
         """Return bounds as tensor.
 
-        Takes bounds from the parameter definitions, but discards bounds belonging to
+        Take bounds from the parameter definitions, but discards bounds belonging to
         columns that were filtered out during the creation of the space.
         """
         if not self.parameters:
@@ -381,10 +391,16 @@ class SubspaceContinuous:
 
     Args:
         parameters: The list of parameters of the subspace.
+        constraints_lin_eq: List of linear equality constraints.
+        constraints_lin_ineq: List of linear inequality constraints.
     """
 
     parameters: List[NumericalContinuousParameter] = field(
         validator=lambda _1, _2, x: _validate_parameter_names(x)
+    )
+    constraints_lin_eq: List[ContinuousLinearEqualityConstraint] = field(factory=list)
+    constraints_lin_ineq: List[ContinuousLinearInequalityConstraint] = field(
+        factory=list
     )
 
     @classmethod
@@ -445,7 +461,7 @@ class SubspaceContinuous:
     def param_bounds_comp(self) -> torch.Tensor:
         """Return bounds as tensor."""
         if not self.parameters:
-            return torch.empty(2, 0)
+            return torch.empty(2, 0, dtype=DTypeFloatTorch)
         return torch.stack([p.bounds.to_tensor() for p in self.parameters]).T
 
     def transform(
@@ -477,9 +493,18 @@ class SubspaceContinuous:
         """
         if not self.parameters:
             return pd.DataFrame()
-        points = torch.distributions.uniform.Uniform(*self.param_bounds_comp).sample(
-            torch.Size((n_points,))
+
+        points = get_polytope_samples(
+            n=n_points,
+            bounds=self.param_bounds_comp,
+            equality_constraints=[
+                c.to_botorch(self.parameters) for c in self.constraints_lin_eq
+            ],
+            inequality_constraints=[
+                c.to_botorch(self.parameters) for c in self.constraints_lin_ineq
+            ],
         )
+
         return pd.DataFrame(points, columns=self.param_names)
 
     def samples_full_factorial(self, n_points: int = 1) -> pd.DataFrame:
@@ -538,9 +563,9 @@ class SearchSpace(SerialMixin):
     continuous: SubspaceContinuous = field(factory=SubspaceContinuous.empty)
 
     def __attrs_post_init__(self):
-        """Validate parameters and record telemetry values."""
+        """Perform validation and record telemetry values."""
         _validate_parameters(self.parameters)
-        _validate_constraints(self.discrete.constraints)
+        _validate_constraints(self.constraints, self.parameters)
 
         # Telemetry
         telemetry_record_value(TELEM_LABELS["COUNT_SEARCHSPACE_CREATION"], 1)
@@ -584,13 +609,17 @@ class SearchSpace(SerialMixin):
         #   ways of object creation) in this particular case.
         _validate_parameters(parameters)
         if constraints:
-            _validate_constraints(constraints)
+            _validate_constraints(constraints, parameters)
+        else:
+            constraints = []
 
         discrete: SubspaceDiscrete = SubspaceDiscrete.from_product(
             parameters=[
                 cast(DiscreteParameter, p) for p in parameters if p.is_discrete
             ],
-            constraints=constraints,
+            constraints=[
+                cast(DiscreteConstraint, c) for c in constraints if c.is_discrete
+            ],
             empty_encoding=empty_encoding,
         )
         continuous: SubspaceContinuous = SubspaceContinuous(
@@ -598,6 +627,16 @@ class SearchSpace(SerialMixin):
                 cast(NumericalContinuousParameter, p)
                 for p in parameters
                 if not p.is_discrete
+            ],
+            constraints_lin_eq=[
+                cast(ContinuousLinearEqualityConstraint, c)
+                for c in constraints
+                if isinstance(c, ContinuousLinearEqualityConstraint)
+            ],
+            constraints_lin_ineq=[
+                cast(ContinuousLinearInequalityConstraint, c)
+                for c in constraints
+                if isinstance(c, ContinuousLinearInequalityConstraint)
             ],
         )
 
@@ -611,7 +650,11 @@ class SearchSpace(SerialMixin):
     @property
     def constraints(self) -> List[Constraint]:
         """Return the constraints of the search space."""
-        return self.discrete.constraints
+        return (
+            self.discrete.constraints
+            + self.continuous.constraints_lin_eq
+            + self.continuous.constraints_lin_ineq
+        )
 
     @property
     def type(self) -> SearchSpaceType: