+++ tests/serialization/test_constraint_serialization.py
@@ -8,7 +8,8 @@ from baybe.constraints import Constraint
 
 @pytest.mark.parametrize(
     "constraint_names",
-    [[f"Constraint_{k}"] for k in range(1, 13)],
+    [[f"Constraint_{k}"] for k in range(1, 13)]  # Constraint_13 is expected to fail
+    + [[f"ContiConstraint_{k}"] for k in range(1, 5)],
 )
 @pytest.mark.parametrize("n_grid_points", [5])
 def test_constraint_serialization(constraints):
@@ -23,9 +24,7 @@ def test_constraint_serialization(constraints):
     [["Constraint_13"]],
 )
 @pytest.mark.parametrize("n_grid_points", [5])
-def test_unsupported_constraints(constraints):
+def test_unsupported_serialization_of_custom_constraint(constraints):
     with pytest.raises(NotImplementedError):
         constraint = constraints[0]
-        string = constraint.to_json()
-        constraint2 = Constraint.from_json(string)
-        assert constraint == constraint2
+        constraint.to_json()