+++ tests/test_constraints_continuous.py
@@ -1,0 +1,116 @@
+"""Test for imposing continuous constraints."""
+import numpy as np
+import pytest
+from baybe.constraints import (
+    ContinuousLinearEqualityConstraint,
+    ContinuousLinearInequalityConstraint,
+)
+
+from .conftest import run_iterations
+
+
+@pytest.mark.parametrize("parameter_names", [["Conti_finite1", "Conti_finite2"]])
+@pytest.mark.parametrize("constraint_names", [["ContiConstraint_1"]])
+@pytest.mark.parametrize("batch_quantity", [5], ids=["b5"])
+def test_equality1(baybe, n_iterations, batch_quantity):
+    """Test equality constraint with equal weights."""
+    run_iterations(baybe, n_iterations, batch_quantity, add_noise=False)
+    res = baybe.measurements_exp
+    print(res)
+
+    assert np.allclose(1.0 * res["Conti_finite1"] + 1.0 * res["Conti_finite2"], 0.3)
+
+
+@pytest.mark.parametrize("parameter_names", [["Conti_finite1", "Conti_finite2"]])
+@pytest.mark.parametrize("constraint_names", [["ContiConstraint_2"]])
+@pytest.mark.parametrize("batch_quantity", [5], ids=["b5"])
+def test_equality2(baybe, n_iterations, batch_quantity):
+    """Test equality constraint with unequal weights."""
+    run_iterations(baybe, n_iterations, batch_quantity, add_noise=False)
+    res = baybe.measurements_exp
+    print(res)
+
+    assert np.allclose(1.0 * res["Conti_finite1"] + 3.0 * res["Conti_finite2"], 0.3)
+
+
+@pytest.mark.parametrize("parameter_names", [["Conti_finite1", "Conti_finite2"]])
+@pytest.mark.parametrize("constraint_names", [["ContiConstraint_3"]])
+@pytest.mark.parametrize("batch_quantity", [5], ids=["b5"])
+def test_inequality1(baybe, n_iterations, batch_quantity):
+    """Test inequality constraint with equal weights."""
+    run_iterations(baybe, n_iterations, batch_quantity, add_noise=False)
+    res = baybe.measurements_exp
+    print(res)
+
+    assert (1.0 * res["Conti_finite1"] + 1.0 * res["Conti_finite2"]).ge(0.299).all()
+
+
+@pytest.mark.parametrize("parameter_names", [["Conti_finite1", "Conti_finite2"]])
+@pytest.mark.parametrize("constraint_names", [["ContiConstraint_4"]])
+@pytest.mark.parametrize("batch_quantity", [5], ids=["b5"])
+def test_inequality2(baybe, n_iterations, batch_quantity):
+    """Test inequality constraint with unequal weights."""
+    run_iterations(baybe, n_iterations, batch_quantity, add_noise=False)
+    res = baybe.measurements_exp
+    print(res)
+
+    assert (1.0 * res["Conti_finite1"] + 3.0 * res["Conti_finite2"]).ge(0.299).all()
+
+
+@pytest.mark.slow
+@pytest.mark.parametrize(
+    "parameter_names",
+    [["Solvent_1", "Conti_finite1", "Conti_finite3", "Conti_finite2"]],
+)
+@pytest.mark.parametrize("constraint_names", [["ContiConstraint_1"]])
+@pytest.mark.parametrize("batch_quantity", [5], ids=["b5"])
+@pytest.mark.parametrize("n_grid_points", [5], ids=["grid5"])
+def test_hybridspace_eq(baybe, n_iterations, batch_quantity):
+    """Test equality constraint with equal weights."""
+    run_iterations(baybe, n_iterations, batch_quantity, add_noise=False)
+    res = baybe.measurements_exp
+    print(res)
+
+    assert np.allclose(1.0 * res["Conti_finite1"] + 1.0 * res["Conti_finite2"], 0.3)
+
+
+@pytest.mark.slow
+@pytest.mark.parametrize(
+    "parameter_names",
+    [["Solvent_1", "Conti_finite1", "Conti_finite3", "Conti_finite2"]],
+)
+@pytest.mark.parametrize("constraint_names", [["ContiConstraint_3"]])
+@pytest.mark.parametrize("batch_quantity", [5], ids=["b5"])
+@pytest.mark.parametrize("n_grid_points", [5], ids=["grid5"])
+def test_hybridspace_ineq(baybe, n_iterations, batch_quantity):
+    """Test inequality constraint with equal weights."""
+    run_iterations(baybe, n_iterations, batch_quantity, add_noise=False)
+    res = baybe.measurements_exp
+    print(res)
+
+    assert (1.0 * res["Conti_finite1"] + 1.0 * res["Conti_finite2"]).ge(0.299).all()
+
+
+def test_invalid_constraints():
+    """Test invalid continuous constraint creations."""
+    # number of parameters and coefficients doesn't match
+
+    with pytest.raises(ValueError):
+        ContinuousLinearEqualityConstraint(
+            parameters=["A", "B"], coefficients=[1.0], rhs=0.0
+        )
+
+    with pytest.raises(ValueError):
+        ContinuousLinearEqualityConstraint(
+            parameters=["A", "B"], coefficients=[1.0, 2.0, 3.0], rhs=0.0
+        )
+
+    with pytest.raises(ValueError):
+        ContinuousLinearInequalityConstraint(
+            parameters=["A", "B"], coefficients=[1.0], rhs=0.0
+        )
+
+    with pytest.raises(ValueError):
+        ContinuousLinearInequalityConstraint(
+            parameters=["A", "B"], coefficients=[1.0, 2.0, 3.0], rhs=0.0
+        )