+++ tests/test_custom_parameter.py
@@ -7,20 +7,10 @@ and metadata. Target objective is minimize to test computational transformation.
 
 import pytest
 
-from baybe.utils.dataframe import add_fake_results, add_parameter_noise
+from .conftest import run_iterations
 
 
 @pytest.mark.parametrize("parameter_names", [["Custom_1", "Custom_2"]])
 def test_run_iterations(baybe, n_iterations, batch_quantity):
     """Test if iterative loop runs with custom parameters."""
-    for _ in range(n_iterations):
-        rec = baybe.recommend(batch_quantity=batch_quantity)
-
-        print(rec)
-        print(baybe.searchspace.discrete.exp_rep)
-        print(baybe.searchspace.discrete.comp_rep)
-
-        add_fake_results(rec, baybe)
-        add_parameter_noise(rec, baybe.parameters, noise_level=0.1)
-
-        baybe.add_measurements(rec)
+    run_iterations(baybe, n_iterations, batch_quantity)