+++ tests/test_substance_parameter.py
@@ -3,7 +3,8 @@
 import pytest
 
 from baybe.utils.chemistry import _MORDRED_INSTALLED, _RDKIT_INSTALLED
-from baybe.utils.dataframe import add_fake_results, add_parameter_noise
+
+from .conftest import run_iterations
 
 _CHEM_INSTALLED = _MORDRED_INSTALLED and _RDKIT_INSTALLED
 if _CHEM_INSTALLED:
@@ -19,11 +20,4 @@ if _CHEM_INSTALLED:
     )
     def test_run_iterations(baybe, batch_quantity, n_iterations):
         """Test running some iterations with fake results and a substance parameter."""
-        for k in range(n_iterations):
-            rec = baybe.recommend(batch_quantity=batch_quantity)
-
-            add_fake_results(rec, baybe)
-            if k % 2:
-                add_parameter_noise(rec, baybe.parameters, noise_level=0.1)
-
-            baybe.add_measurements(rec)
+        run_iterations(baybe, n_iterations, batch_quantity)